# Faster UTF16 -> UTF8 and UTF8 -> UTF16