… src/bun.js/bindings/bun-simdutf.zig
… src/bun.js/bindings/simdutf.cpp
… src/bun.js/bindings/simdutf.h