This adds simdutf as a dependency to Bun and uses it in some of the string handling functions to and from JavaScript

### writeFileSync

~1.75x faster for non-ascii text

After:

```zig
cpu: Apple M1 Max
runtime: bun 0.3.0 (arm64-darwin)

benchmark        time (avg)             (min … max)       p75       p99      p995
--------------------------------------------------- -----------------------------
12 ascii      19.14 µs/iter  (13.21 µs … 100.33 µs)  21.29 µs  30.04 µs  36.33 µs
12 utf8       17.95 µs/iter   (13.13 µs … 98.58 µs)  18.54 µs  25.92 µs  29.08 µs
12288 ascii   20.98 µs/iter  (16.71 µs … 100.13 µs)  22.21 µs     32 µs  35.67 µs
18432 utf8     38.4 µs/iter  (32.46 µs … 187.17 µs)  41.33 µs  53.88 µs     61 µs
```

Before:

```zig
cpu: Apple M1 Max
runtime: bun 0.3.0 (arm64-darwin)

benchmark        time (avg)             (min … max)       p75       p99      p995
--------------------------------------------------- -----------------------------
12 ascii      18.07 µs/iter    (13.08 µs … 1.31 ms)  18.08 µs     29 µs  35.38 µs
12 utf8       18.12 µs/iter   (13.21 µs … 98.38 µs)  18.79 µs  26.13 µs  29.88 µs
12288 ascii   21.22 µs/iter   (16.54 µs … 94.46 µs)   22.5 µs  32.04 µs  36.83 µs
18432 utf8    67.97 µs/iter  (62.88 µs … 233.75 µs)  68.88 µs  79.08 µs  83.25 µs
```


### `TextDecoder` and `Buffer.toString`

~1.6x  faster for non-ascii text

After:
```zig
cpu: Apple M1 Max
runtime: bun 0.3.0 (arm64-darwin)

benchmark               time (avg)             (min … max)       p75       p99      p995
---------------------------------------------------------- -----------------------------
12 ascii             55.05 ns/iter  (50.27 ns … 402.54 ns)  53.55 ns 106.05 ns 124.82 ns
12 utf8             100.12 ns/iter  (93.99 ns … 247.79 ns)  99.29 ns 154.62 ns  160.5 ns
12288 ascii         836.24 ns/iter   (748.63 ns … 1.34 µs) 843.43 ns   1.34 µs   1.34 µs
24576 utf8           16.58 µs/iter    (15.33 µs … 1.71 ms)     16 µs  19.04 µs   21.5 µs
Buffer 12 ascii      65.27 ns/iter  (57.07 ns … 139.69 ns)  66.48 ns  73.25 ns 112.76 ns
Buffer 12 utf8       53.55 ns/iter  (49.16 ns … 140.73 ns)  52.81 ns  63.15 ns   65.2 ns
Buffer 12288 ascii  717.18 ns/iter (639.01 ns … 867.13 ns) 735.82 ns 867.13 ns 867.13 ns
Buffer 24576 utf8    16.48 µs/iter    (15.29 µs … 1.83 ms)     16 µs  19.29 µs  20.25 µs
```

Before:

```zig
cpu: Apple M1 Max
runtime: bun 0.3.0 (arm64-darwin)

benchmark               time (avg)             (min … max)       p75       p99      p995
---------------------------------------------------------- -----------------------------
12 ascii             38.99 ns/iter  (35.56 ns … 197.91 ns)   37.9 ns  90.49 ns  98.71 ns
12 utf8              92.74 ns/iter   (89.2 ns … 186.41 ns)  91.88 ns 143.99 ns 167.94 ns
12288 ascii           1.08 µs/iter   (955.07 ns … 1.61 µs)   1.11 µs   1.61 µs   1.61 µs
24576 utf8           26.06 µs/iter  (24.17 µs … 921.38 µs)  25.46 µs  30.17 µs  34.71 µs
Buffer 12 ascii       62.7 ns/iter   (54.5 ns … 127.82 ns)  64.24 ns  69.34 ns  72.66 ns
Buffer 12 utf8        40.6 ns/iter  (36.62 ns … 101.24 ns)     40 ns  52.05 ns  60.83 ns
Buffer 12288 ascii  774.37 ns/iter   (683.63 ns … 1.11 µs) 804.83 ns   1.11 µs   1.11 µs
Buffer 24576 utf8    25.38 µs/iter    (24.21 µs … 1.41 ms)  25.08 µs  27.83 µs  29.79 µs
```

### TextEncoder

~2.8x faster for UTF8 -> UTF16

Before:

```zig
cpu: Apple M1 Max
runtime: bun 0.3.0 (arm64-darwin)

benchmark        time (avg)             (min … max)       p75       p99      p995
--------------------------------------------------- -----------------------------
4 ascii       21.56 ns/iter  (20.23 ns … 136.42 ns)   20.9 ns  55.41 ns  65.08 ns
4 utf8        30.76 ns/iter  (29.11 ns … 127.53 ns)  30.14 ns  65.09 ns  74.88 ns
12 ascii      24.38 ns/iter   (22.68 ns … 452.2 ns)  23.44 ns  66.23 ns  75.36 ns
12 utf8       45.88 ns/iter   (43.3 ns … 106.79 ns)  44.92 ns  89.27 ns   95.1 ns
12288 ascii   11.08 µs/iter    (10.21 µs … 1.35 ms)  10.88 µs  12.33 µs  13.13 µs
18432 utf8    38.96 µs/iter    (36.58 µs … 2.13 ms)  38.13 µs  42.04 µs   44.5 µs
```


After:

```zig
cpu: Apple M1 Max
runtime: bun 0.3.0 (arm64-darwin)

benchmark        time (avg)             (min … max)       p75       p99      p995
--------------------------------------------------- -----------------------------
4 ascii       21.57 ns/iter  (19.93 ns … 144.52 ns)   20.6 ns  59.92 ns  73.25 ns
4 utf8         36.2 ns/iter  (32.89 ns … 253.77 ns)  36.06 ns  82.54 ns 118.08 ns
12 ascii      24.85 ns/iter   (22.7 ns … 153.99 ns)  24.34 ns   66.8 ns  84.44 ns
12 utf8       47.67 ns/iter  (45.13 ns … 166.38 ns)  46.58 ns  92.15 ns 109.33 ns
12288 ascii   11.09 µs/iter  (10.25 µs … 812.13 µs)  10.88 µs  12.25 µs  13.25 µs
18432 utf8    13.51 µs/iter    (12.42 µs … 2.18 ms)  12.96 µs  15.21 µs  15.67 µs
```