@Jarred-Sumner 2022-11-27
cc @lemire this is using [simdutf](https://github.com/simdutf/simdutf/)